refactor: 인풋창 불필요한 리렌더링을 방지하기 위해 컴포넌트를 분리하라
- 검색 창 인풋창 상태관리로 인해 리렌더링 발생하여 컴포넌트화 시켜 리렌더링 방지
