+++ components/map/SampleMap.tsx
@@ -3,22 +3,11 @@ import {
 } from 'react';
 
 import { GoogleMap, StandaloneSearchBox, useLoadScript } from '@react-google-maps/api';
-import useActionKeyEvent from 'hooks/useActionKeyEvent';
 import { PlaceGeometry } from 'lib/types/google.maps';
-import { styled } from 'styled-components';
-import { headlineFont } from 'styles/fontStyles';
 
 import PlaceBottomSheet from './PlaceBottomSheet';
 import PlaceResultMarker from './PlaceResultMarker';
-
-import CloseIcon from 'lib/assets/icons/close.svg';
-import MenuIcon from 'lib/assets/icons/menu.svg';
-import SearchIcon from 'lib/assets/icons/search.svg';
-
-const center = {
-  lat: 0,
-  lng: -180,
-};
+import SearchInput from './SearchInput';
 
 // TODO - Sample Map 추후 리팩터링 및 수정
 function SampleMap() {
@@ -33,10 +22,6 @@ function SampleMap() {
   const [searchBoxState, setSearchBoxState] = useState<google.maps.places.SearchBox>();
   const [placeResults, setPlaceResult] = useState<PlaceGeometry[]>([]);
   const [mapState, setMapState] = useState<google.maps.Map | null>(null);
-  const [searchInput, setSearchInput] = useState<string>('');
-
-  const onKeyDown = useActionKeyEvent<HTMLInputElement>(['Enter', 'NumpadEnter'], (e) => setSearchInput(e.currentTarget.value));
-  const onDeleteInput = () => setSearchInput('');
 
   const onLoad = useCallback((map: google.maps.Map) => {
     setMapState(map);
@@ -103,8 +88,8 @@ function SampleMap() {
           width: '100%',
           maxWidth: '430px',
         }}
-        zoom={2}
-        center={center}
+        zoom={3}
+        center={new google.maps.LatLng(36.204824, 138.252924)}
         onUnmount={onUnmount}
         onLoad={onLoad}
         onBoundsChanged={() => {
@@ -115,31 +100,16 @@ function SampleMap() {
         }}
       >
         <StandaloneSearchBox
-          bounds={google.maps.LatLngBounds.MAX_BOUNDS}
+          // NOTE - 일본 bounds
+          bounds={new google.maps.LatLngBounds(
+            new google.maps.LatLng(20.3585295, 122.8554688),
+            new google.maps.LatLng(45.6412626, 154.0031455),
+          )}
           onLoad={onLoadSearchBox}
           onPlacesChanged={onPlacesChanged}
           onUnmount={() => setSearchBoxState(undefined)}
         >
-          {/* TODO - 인풋 컴포넌트 추후 공통 컴포넌트로 변경 */}
-          <InputWrapper>
-            <SearchIconWrapper>
-              <SearchIcon className="search-icon" />
-            </SearchIconWrapper>
-            <Input
-              type="text"
-              placeholder="장소 검색"
-              value={searchInput}
-              onChange={(e) => setSearchInput(e.target.value)}
-              onKeyDown={onKeyDown}
-            />
-            <MenuIconWrapper>
-              {searchInput.trim() ? (
-                <CloseIcon onClick={onDeleteInput} />
-              ) : (
-                <MenuIcon />
-              )}
-            </MenuIconWrapper>
-          </InputWrapper>
+          <SearchInput />
         </StandaloneSearchBox>
         {placeResults.map((place) => (
           <PlaceResultMarker key={place.place_id} place={place} />
@@ -152,61 +122,3 @@ function SampleMap() {
 }
 
 export default memo(SampleMap);
-
-const SearchIconWrapper = styled.div`
-  cursor: pointer;
-  position: absolute;
-  display: flex;
-  top: 16px;
-  left: 20px;
-`;
-
-const MenuIconWrapper = styled.div`
-  display: flex;
-  position: absolute;
-  top: 16px;
-  right: 20px;
-  cursor: pointer;
-`;
-
-const InputWrapper = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  top: 24px;
-  right: 24px;
-  position: absolute;
-  width: calc(100% - 48px);
-
-  &:focus-within {
-    .search-icon > path {
-      fill: ${({ theme }) => theme.purple500};
-    }
-  }
-`;
-
-const Input = styled.input`
-  ${headlineFont({ fontWeight: 500 })};
-  letter-spacing: -0.012em;
-
-  color: ${({ theme }) => theme.black};
-  background: ${({ theme }) => theme.white};
-  box-shadow: 0px 4px 0px rgba(19, 17, 24, 0.16);
-  outline: none;
-  width: 100%;
-  height: 56px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 12px 60px;
-  border: 2px solid ${({ theme }) => theme.black};
-  border-radius: 56px;
-
-  &::placeholder {
-    color: ${({ theme }) => theme.gray400};
-  }
-
-  &:focus {
-    caret-color: ${({ theme }) => theme.purple500};
-  }
-`;
